Shift geofield type to model file
